attaches display map to details button
