+++ public/index.html
@@ -25,9 +25,9 @@
   <script type="text/javascript" src="/scripts/networks.js"></script>
 <!-- Component Stuff -->
   <script type="text/javascript" src="/scripts/components/index.js"></script>
+  <script type="text/javascript" src="/scripts/components/map.js"></script>
   <script type="text/javascript" src="/scripts/components/map_listing.js"></script>
   <script type="text/javascript" src="/scripts/components/map_listings.js"></script>
-  <script type="text/javascript" src="/scripts/components/map.js"></script>
   <script type="text/javascript" src="/scripts/components/registerform.js"></script>
   <script type="text/javascript" src="/scripts/components/loginform.js"></script>
   <script type="text/javascript" src="/scripts/components/header.js"></script>
